chore: move all wasm gating in mod.rs
